Implement compression
Add compression support to the library. For now this is a bit too slow
to be usable, and the ratio is not as good as Apple's because I haven't
implemented the repetition counts. I plan to fix the second issue soon,
but the first one may not be as easy.