Don't confuse trailing bitmaps with a repetition
My tests report an error when trying to decompress the result of
compressing 006-original. What happens is that the decompression
mistakes a trailing bitmap that starts with 0xf with a repetition count
for the preceding bitmap number.

I've confirmed that Apple's library correctly parses my compressed file,
so the bug is in the decompression. Check if we are done with the chunk
before doing any work in zbm_read_repetition_count().

To make sure that we keep test coverage for this issue even if the
compression code changes, rig my current compressed output for
006-original as the new 019 test.