Check the magic one byte at a time
The latest patch got rid of most multibyte ints on the file layout. The
only one that remains is the magic, which is handled as an uint32_t.
Make that a string so that I can claim to support big-endian
architectures, which I guess might matter somewhat for the kernel
driver.